update curve interface

+++ src/interfaces/ICurveFi.sol
@@ -22,9 +22,15 @@ interface ICurveFi {
         uint256 min_mint_amount
     ) external payable;
 
-    function remove_liquidity_imbalance(uint256[4] calldata amounts, uint256 max_burn_amount) external;
+    function remove_liquidity_imbalance(
+        uint256[4] calldata amounts,
+        uint256 max_burn_amount
+    ) external;
 
-    function remove_liquidity(uint256 _amount, uint256[4] calldata amounts) external;
+    function remove_liquidity(
+        uint256 _amount,
+        uint256[4] calldata amounts
+    ) external;
 
     function remove_liquidity_one_coin(
         uint256 _token_amount,
@@ -42,18 +48,17 @@ interface ICurveFi {
     function balances(int128) external view returns (uint256);
 
     function get_dy(
-        int128 from,
-        int128 to,
+        uint256 from,
+        uint256 to,
         uint256 _from_amount
     ) external view returns (uint256);
 
-    function calc_token_amount( uint256[2] calldata amounts, bool is_deposit) external view returns (uint256);
+    function calc_token_amount(
+        uint256[2] calldata amounts,
+        bool is_deposit
+    ) external view returns (uint256);
 }
 
 interface Zap {
-    function remove_liquidity_one_coin(
-        uint256,
-        int128,
-        uint256
-    ) external;
+    function remove_liquidity_one_coin(uint256, int128, uint256) external;
 }
